backend/requirements.txt
Django==4.2.4
djangorestframework==3.14.0
django-cors-headers==4.2.0
mysqlclient==2.2.0
python-dotenv==1.0.0
django-filter==23.2
Pillow==10.0.0
PyJWT==2.8.0